Switched to the crystal programming language
I switched to the crystal programming language because it is faster and
it produces binaries. In order to switch I edited flake.nix and many
small parts of what was xtodoc.rb, and is now xtodoc.cr. I also began to
add multithreading to xtodoc as preparation for the change to a more
extensible design (I know extensible design means nothing, I have not yet
nailed down the architecture).