Rearranged files, changed how Pandoc is integrated
Rearranged files so that the theme/css and all the other per site
stylization is integrated in such a way that the path for this project's
stylization is the same relative path will be used by projects with
customized styling. For now there is no default stylization, meaning
every site must provide their own, this will change.

Change how Pandoc is integrated. This method seems to be more reliable.